♻️ refactor: 소통게시판 목록 API apiClient로 교체
