+++ src/app/(routers)/(board)/community/_api/communityQueries.ts
@@ -1,0 +1,24 @@
+import type { Post, SortOption } from '../types';
+
+import { apiClient } from '@/lib/apiClient';
+import { useQuery } from '@tanstack/react-query';
+
+import { communityQueryKeys } from './communityQueryKeys';
+
+export interface PostsResponse {
+  posts: Post[];
+  nextCursor: string | null;
+  totalCount: number;
+}
+
+const toApiType = (sort: SortOption): 'all' | 'best' => (sort === '인기순' ? 'best' : 'all');
+
+// 게시물 전체 조회
+export const useGetPosts = (sort: SortOption = '최신순') => {
+  const type = toApiType(sort);
+  return useQuery({
+    queryKey: communityQueryKeys.posts(type),
+    queryFn: () => apiClient<PostsResponse>(`/posts?type=${type}`),
+    staleTime: 1000 * 60 * 5,
+  });
+};
