+++ src/app/(routers)/(board)/community/_api/communityQueryKeys.ts
@@ -1,4 +1,4 @@
 export const communityQueryKeys = {
   all: ['community'] as const,
-  posts: () => [...communityQueryKeys.all, 'posts'] as const,
+  posts: (type: 'all' | 'best' = 'all') => [...communityQueryKeys.all, 'posts', type] as const,
 } as const;
