+++ src/app/(routers)/(board)/community/page.tsx
@@ -1,20 +1,5 @@
-import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
-
-import { getPosts } from './_api/communityApi';
-import { communityQueryKeys } from './_api/communityQueryKeys';
 import CommunityClient from './CommunityClient';
 
-export default async function CommunityPage() {
-  const queryClient = new QueryClient();
-
-  await queryClient.prefetchQuery({
-    queryKey: communityQueryKeys.posts(),
-    queryFn: getPosts,
-  });
-
-  return (
-    <HydrationBoundary state={dehydrate(queryClient)}>
-      <CommunityClient />
-    </HydrationBoundary>
-  );
+export default function CommunityPage() {
+  return <CommunityClient />;
 }
